last updates with all improvements
